% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_get.R
\name{rc_get}
\alias{rc_get}
\title{Retrieve a MP votes on a single roll call vote}
\usage{
rc_get(voteid = NA, include_voteinfo = FALSE, include_mpinfo = FALSE)
}
\value{
A data.frame with the following variables:\tabular{ll}{
    \tab  \cr
   \strong{response_date} \tab Date of data retrieval \cr
   \strong{version} \tab Data version from the API \cr
   \strong{vote_id} \tab Id of vote \cr
   \strong{mp_id} \tab MP id \cr
   \strong{party_id} \tab Party id \cr
   \strong{vote} \tab Vote: for, mot (against), ikke_tilstede (absent) \cr
   \strong{permanent_sub_for} \tab Id of the MP originally holding the seat, if the substitute is permanent \cr
   \strong{sub_for} \tab Id of the MP originally holding the seat \cr
}
}
\description{
A function for retrieving which MPs voted what on a single vote
in the Norwegian parliament.
}
\examples{
\dontrun{

v <- rc_get(12345)

}

}
\seealso{
\code{\link[stortingscrape:get_vote]{stortingscrape::get_vote()}}
}
