% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_get.R
\name{rc_get}
\alias{rc_get}
\title{Retrieve a MP votes on a single roll call vote}
\usage{
rc_get(voteid = NA, include_voteinfo = FALSE, include_mpinfo = FALSE)
}
\arguments{
\item{voteid}{Integer. One or more vote ids to retreive from the API.}

\item{include_voteinfo}{Logical. Whether or not to include information on the vote.}

\item{include_mpinfo}{Logical. Wheteher or not to include information on the MPs.}

\item{good_manners}{Numeric. Seconds to wait between calls to the API.
Should be used if \code{voteid} is a vector longer than 1.}

\item{...}{Other arguments passed to \code{rollcall()}}
}
\value{
A rollcall object with the specified votes
}
\description{
A function for retrieving which MPs voted what on a single vote
in the Norwegian parliament.
}
\examples{
\dontrun{

v <- rc_get(12345)
summary(v, verbose = TRUE)

}

}
\seealso{
\code{\link[stortingscrape:get_vote]{stortingscrape::get_vote()}}
}
